refactor: Sum Layout Files
