+++ FE/src/app/layout.jsx
@@ -1,0 +1,5 @@
+import RootLayout from "@/components/common/Layout";
+
+export default function Layout({ children }) {
+  return <RootLayout>{children}</RootLayout>;
+}
